feat: adiciona DTOs de requisição para contas e transações

+++ src/main/java/br/com/bytebank/dto/AbrirContaRequest.java
@@ -6,12 +6,14 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.Pattern;
 
+import java.math.BigDecimal;
+
 public record AbrirContaRequest(
 
         @Nonnull Long clienteId,
         @Nonnull int numero,
 
-        @Nonnull @Positive double saldoInicial,
+        @Nonnull @Positive BigDecimal saldoInicial,
 
 
         @Nonnull @Pattern(regexp = "CC|CP", message = "O tipo de conta deve ser CC (Corrente) ou CP (Poupança)") //diferenciando dentro da tabela conta
