+++ src/main/java/br/com/bytebank/dto/TransacaoRequest.java
@@ -3,10 +3,12 @@ package br.com.bytebank.dto;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
+
 public record TransacaoRequest(
         @NotNull
         @Positive
-        double valor
+        BigDecimal valor
 
 ) {
 }
