+++ src/main/java/br/com/bytebank/dto/TransferenciaRequest.java
@@ -3,11 +3,13 @@ package br.com.bytebank.dto;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
+
 public record TransferenciaRequest(
 
         @NotNull
         @Positive
-        double valor,
+        BigDecimal valor,
 
         @NotNull
         Long contaDestinoId
